app/requirements.txt
Django==4.2.5
gunicorn==21.2.0
psycopg2-binary==2.9.7

djangorestframework==3.14.0
drf-spectacular==0.26.4
Faker==19.6.0
django-seed==0.3.1
boto3==1.28.49
